+++ src/modules/esm.mjs
@@ -1,0 +1,46 @@
+import path from "path";
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import output from "./files/c.mjs";
+import { createRequire } from "module";
+import { fileURLToPath } from "url";
+
+output();
+
+const random = Math.random();
+
+let unknownObject;
+
+const require = createRequire(import.meta.url);
+if (random > 0.5) {
+    unknownObject = require("./files/a.json");
+} else {
+    unknownObject = require("./files/b.json");
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+const __filename = fileURLToPath(import.meta.url);
+console.log(`Path to current file is ${__filename}`);
+const __dirname = path.dirname(__filename);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log("To terminate it, use Ctrl+C combination");
+});
+
+export default {
+    unknownObject,
+    myServer,
+};
