+++ src/modules/files/c.mjs
@@ -1,0 +1,3 @@
+export default function output() {
+    console.log("Hello from c.js!");
+}
